pyproject.toml
[project]
name = "cyberark-tpc-plugin-validator"
version = "0.9.0"
description = "A tool to help validate a custom TPC plugin."
keywords = ["cyberark", "validate", "tpc"]
authors = [{ name = "Peter McDonald", email = "git@petermcdonald.co.uk"}]
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.11"
dependencies = [
    "cyberark-tpc-plugin-parser>=0.3.1",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Programming Language :: Python :: 3.15",
    "Topic :: Software Development :: Testing",
    "Typing :: Typed",
]

[project.urls]
homepage = "https://github.com/petermcd/CyberArk-TPC-Plugin-Validator"
repository = "https://github.com/petermcd/CyberArk-TPC-Plugin-Validator.git"
issues = "https://github.com/petermcd/CyberArk-TPC-Plugin-Validator/issues"

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pytest-cov>=6.3.0",
    "pytest-mypy-plugins>=3.2.0",
    "ruff>=0.12.12",
    "troml>=0.4.1",
    "ty>=0.0.1a20",
]

[project.scripts]
tpc-validator = "tpc_plugin_validator.main:main"

[tool.pytest.ini_options]
addopts = "--cov=tpc_plugin_validator --cov-report term-missing --cov-report xml:coverage.xml"

[tool.ruff]
line-length = 120
